conversational-workflow/requirements.txt
langgraph>=0.2.28
langchain>=0.1.0
langchain-core>=0.2.39
langchain-openai>=0.1.0
openai>=1.10.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
